Remove FileWriterLCIOC and adapt the more general file FileWriterLCIO. Add timestamp in the header for all CALICE lcio events (needed to acces to the calibration data base
